quantum_llm_waf/secure_agent/requirements.txt
langgraph
langchain-openai
langchain-core
requests
python-dotenv
termcolor
tqdm
transformers
torch
accelerate
